Download imgs, fix @media css, add html img alt
Two new photos for 'random info' section. Still need two more, but
reached daily download limit. Do it later.

On mobile viewport, the page background cut off on the right side with
the edge of the navbar and header going through the img. Push and check
if @media adjustment to flex column fixed it.